Move FlyView to abby
