+++ src/main/java/examples/collision/ConeCollision.java
@@ -8,14 +8,14 @@ import javax.media.opengl.GL2;
 import com.badlogic.gdx.math.Intersector;
 import com.badlogic.gdx.math.collision.Ray;
 
-import examples.simple.StandardExample;
+import br.com.abby.core.view.FlyView;
 import br.com.etyllica.core.event.KeyEvent;
 import br.com.etyllica.core.event.PointerEvent;
 import br.com.luvia.core.context.ApplicationGL;
-import br.com.luvia.core.controller.FlyView;
 import br.com.luvia.core.graphics.Graphics3D;
 import br.com.luvia.graphics.Cone;
 import br.com.luvia.graphics.Cylinder;
+import examples.simple.StandardExample;
 
 public class ConeCollision extends ApplicationGL {
